# api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: Odal Node API
  version: 0.12.0
  description: |
    **Sovereign Digital Product Passport Infrastructure**

    The Odal API is split across two deployable services:

    - **Node** (`http://localhost:8001`) — the MVP service that bundles vault,
      identity, and integrator on a single port with sub-path routing:
      `/vault/*`, `/identity/*`, `/integrator/*`.
    - **Resolver** (`http://localhost:8003`) — standalone public resolver for
      QR code scans and supply-chain integrations. Unauthenticated.

    All authenticated vault endpoints are versioned under `/vault/api/v1/`.
    Path parameters use OpenAPI 3.1 `{id}` brace syntax (aligned with
    Axum 0.8 routing).

    Authentication uses `Authorization: Bearer odal_sk_...` API keys (SHA-256
    hashed, prefix-indexed) or HTTP Basic auth for local development.

    Regulation references:
    - EU Ecodesign for Sustainable Products Regulation (ESPR) 2024/1781
    - EU Battery Regulation 2023/1542
    - W3C DID Core Spec 1.0 / W3C VC Data Model 2.0
  contact:
    name: Odal Support
    email: contact@odal-node.io
    url: https://docs.odal-node.io
  license:
    name: BSL-1.1
    url: https://mariadb.com/bsl11/
  x-legal-entity: Odal Node
servers:
  - url: http://localhost:8001
    description: Local development — odal-node (vault + identity public + integrator)
  - url: http://localhost:8002
...
  - url: http://localhost:8003
    description: Local development — odal-resolver (public)
...
paths:
...
  /internal/sign:
    servers:
      - url: http://localhost:8002
        description: odal-identity standalone (mTLS internal)
    post:
      operationId: internalSign
      summary: Sign a payload (internal, mTLS)
      description: |
        Signs a base64-encoded canonical-JSON payload with the operator's
        Ed25519 key (auto-provisioned on first use). Returns a compact JWS.
        Service-to-service only — gated by mTLS (`CN=odal-vault`).
      tags:
        - Identity (internal)
      security:
        - MutualTLS: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignRequest'
      responses:
        '200':
          description: Signed payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignResponse'
        '400':
          description: Payload is not valid base64 or not valid JSON.
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/verify:
    servers:
      - url: http://localhost:8002
        description: odal-identity standalone (mTLS internal)
    post:
      operationId: internalVerify
      summary: Verify a JWS this service issued (internal, mTLS)
      description: |
        Checks a compact JWS against the named operator's key *and* confirms
        it was signed over the given payload — a validly-signed JWS for
        different content does not pass. Never errors on a signature that
        simply fails to verify; that is `{ "valid": false }`, not a fault.
        Service-to-service only — gated by mTLS (`CN=odal-vault`).
      tags:
        - Identity (internal)
      security:
        - MutualTLS: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: 'Verification result. Always `200` — an unverifiable signature is `{ "valid": false }`, not an error status.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: |
        Machine-to-machine access via API keys. Keys are prefixed with
        `odal_sk_` and sent as `Authorization: Bearer odal_sk_...`.

        Keys are stored as SHA-256 hashes in PostgreSQL with a 12-character
        prefix index for fast lookup. The plain-text key is only returned
        once at creation time via `POST /vault/api/v1/api-keys`.
    BasicAuth:
      type: http
      scheme: basic
      description: |
        The operator's own bootstrap credential, from the `ADMIN_USERNAME` and
        `ADMIN_PASSWORD` environment variables. It mints the first API key on a
        fresh node — before any key exists — and is the lockout-recovery path
        afterwards, since it carries no `keyId` and so can revoke any key.

        Active in every environment where both variables are set; there is no
        production gate. Leave them unset once an API key exists if you do not
        want the path available.

        Reached only via the `Basic` scheme. A `Bearer` token is never matched
        against it, even one carrying the same `base64(user:pass)` payload.
    MutualTLS:
      type: mutualTLS
      description: |
        Client-certificate (mTLS) authentication for service-to-service calls.
        The standalone identity service accepts the internal `/internal/*`
        endpoints only from a client presenting `CN=odal-vault`. The fused
        `dpp-node` never exposes these over the network — it signs in-process.
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 / RFC 9457 problem details. The shape
        `dpp-common::http_problem::Problem` produces, served as
        `application/problem+json`.

        `type` is derived from `title`, so each distinct `title` used across the
        codebase is a stable catalogue key that clients may depend on.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Absolute URI identifying the problem type.
          example: https://problems.odal-node.io/not-found
        title:
          type: string
          description: Short human-readable summary of the problem type.
          example: Not Found
        status:
          type: integer
          description: The HTTP status code, mirroring the status line.
          example: 404
        detail:
          type: string
          description: Human-readable explanation for this specific occurrence.
          example: 'No schema for product group ''nosuchsector''. Known product groups: aluminium, battery.'
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this specific occurrence.
...
    SignRequest:
      type: object
      required:
        - operator_id
        - passport_id
        - payload
      description: Internal signing request. Field names are snake_case (internal contract).
      properties:
        operator_id:
          type: string
          description: Operator whose key signs. Auto-provisioned on first use.
          example: self_hosted
        passport_id:
          type: string
          description: The passport id being signed (recorded in the JWS payload).
        payload:
          type: string
          description: Base64-encoded canonical JSON of the payload to sign.
    SignResponse:
      type: object
      required:
        - jws_signature
      properties:
        jws_signature:
          type: string
          description: Compact JWS (EdDSA over RFC 8785 canonical bytes).
    VerifyRequest:
      type: object
      required:
        - operator_id
        - jws
        - payload
      description: Internal verification request. Field names are snake_case (internal contract).
      properties:
        operator_id:
          type: string
          description: Operator id whose key the signature is checked against.
          example: self_hosted
        jws:
          type: string
          description: The compact JWS to verify.
        payload:
          description: The payload the caller expects the JWS to have been signed over.
    VerifyResponse:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
          description: True iff the signature verifies against the named operator's key AND was signed over exactly this payload.
...
  responses:
...
    ValidationError:
      description: One or more fields failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://problems.odal-node.io/unprocessable-entity
            title: Unprocessable Entity
            status: 422
            detail: productName must not be empty.
...